fix(feature-selection): widen sample-count products in rank, contingency and point-biserial stats
CodeQL cs/loss-of-precision in the Kruskal-Wallis, Wilcoxon, Fisher exact, adjusted mutual
information, point-biserial and bimodality selectors: sample-count products were computed in
int before being used as double. Thresholds where results go wrong: Wilcoxon signed-rank
n(n+1)(2n+1) at 1,024 pairs (std NaN, p = 1); Kruskal-Wallis tie correction count^3 at 1,291;
n(n+1), n0*n1, a*d, b*c, ai*bj, (n-2)(n-3) at ~46k-93k samples. FisherExactTest also tested
`b * c > 0` in int, which can flip sign; it now tests `b > 0 && c > 0`.

+++ src/Preprocessing/FeatureSelection/Filter/Statistical/WilcoxonSignedRank.cs
@@ -121,8 +121,8 @@ public class WilcoxonSignedRank<T> : TransformerBase<T, Matrix<T>, Matrix<T>>
 
             // Normal approximation for p-value
             int nPairs = differences.Count;
-            double mean = nPairs * (nPairs + 1) / 4.0;
-            double stdDev = Math.Sqrt(nPairs * (nPairs + 1) * (2 * nPairs + 1) / 24.0);
+            double mean = (double)nPairs * (nPairs + 1) / 4.0;
+            double stdDev = Math.Sqrt((double)nPairs * (nPairs + 1) * (2 * nPairs + 1) / 24.0);
 
             if (stdDev > 0)
             {
